Adjusted deleteAll so that admin accounts is not deleted

+++ src/models/mongo/user-mongo-store.js
@@ -46,6 +46,6 @@ export const userMongoStore = {
   },
 
   async deleteAll() {
-    await User.deleteMany({});
+    await User.deleteMany({email: {$ne:"admin@admin.com"}});
   }
 };
